Repo // Moving input handling instructions to InputHandler.
